Fix type mismatch in PostgreSQL bigserial integration test by casting InsertGetId return value to int64
- Cast id from InsertGetId to int64 when comparing with user.ID in TestPostgreSQLIntegrationQueryInsertGetIdBigSerial
- Resolves type comparison error between int and int64 for bigserial primary key validation

+++ integration_tests/postgres/postgres_query_create_test.go
@@ -139,7 +139,7 @@ func TestPostgreSQLIntegrationQueryInsertGetIdBigSerial(t *testing.T) {
 	if id == 0 {
 		t.Error("ID should not be zero for bigserial")
 	}
-	if user.ID != id {
+	if user.ID != int64(id) {
 		t.Errorf("Expected ID %d, got %d", id, user.ID)
 	}
